requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Session==0.8.0
PyMySQL==1.1.1
openpyxl==3.1.5
python-dotenv==1.0.1
